+++ packages/gitbook/src/components/primitives/Link.tsx
@@ -4,6 +4,7 @@ import NextLink, { type LinkProps as NextLinkProps } from 'next/link';
 import React from 'react';
 
 import { tcls } from '@/lib/tailwind';
+import { checkIsAnchor, resolveAnchorURL } from '@/lib/urls';
 import { type TrackEventInput, useTrackEvent } from '../Insights';
 import { NavigationStatusContext } from '../hooks';
 import { isExternalLink, toNonEmbedLink } from '../utils/link';
@@ -58,6 +59,10 @@ function getTargetProps(
         isExternal: boolean;
     }
 ) {
+    if (props.href.startsWith('mailto:')) {
+        return {};
+    }
+
     const target =
         props.target ??
         (context.isExternal && context.externalTarget === '_blank' ? '_blank' : undefined);
@@ -88,12 +93,20 @@ export function Link(props: LinkProps) {
     const trackEvent = useTrackEvent();
     const forwardedClassNames = useClassnames(classNames || []);
     const isExternal = isExternalServer(href);
+    const isAnchor = checkIsAnchor(href);
     const { target, rel } = getTargetProps(props, { externalTarget, isExternal });
 
     const onClick = (event: React.MouseEvent<HTMLAnchorElement>) => {
         // Only trigger navigation context for internal links in the same window without modifier keys (i.e. open in new tab).
         if (!isExternal && target !== '_blank' && !event.ctrlKey && !event.metaKey) {
-            onNavigationClick(href);
+            if (isAnchor) {
+                event.preventDefault();
+                const resolvedHref = resolveAnchorURL(href, window.location);
+                window.history.pushState(null, '', resolvedHref);
+                onNavigationClick(resolvedHref);
+            } else {
+                onNavigationClick(href);
+            }
         }
 
         const isExternalOnClient = isExternalClient(href);