+++ packages/gitbook/src/intl/translations/ru.ts
@@ -9,6 +9,7 @@ export const ru = {
     switch_to_system_theme: 'Переключиться на системную тему',
     search: 'Поиск',
     clear: 'Очистить',
+    tags: 'Теги',
     search_back: 'Вернуться к результатам поиска',
     search_or_ask: 'Найти или спросить',
     search_input_placeholder: 'Поиск по содержимому',
@@ -74,6 +75,7 @@ export const ru = {
     notfound_adaptive:
         'Эта страница может существовать, но для доступа к ней может потребоваться вход.',
     notfound_adaptive_login: 'Войти',
+    notfound_goto_home: 'Перейти на главную',
     unexpected_error_title: 'Произошла ошибка',
     unexpected_error: 'Извините, произошла непредвиденная ошибка. Пожалуйста, попробуйте позже.',
     unexpected_error_retry: 'Попробуйте снова',
@@ -163,4 +165,5 @@ export const ru = {
     form_other_prompt: 'Введите ответ...',
     form_other_field: 'Другое...',
     cancel: 'Отмена',
+    direct_link_to_heading: 'Прямая ссылка на заголовок',
 };